+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-04-10 15:59:26
-// Commit: test(state): implement widget integration tests for state
+// Updated at: 2026-04-10 20:14:48
+// Commit: style(state): harmonize light and dark theme palette for state
 
 class stateComponentState {
-  final String commitRef = '282';
-  final String lastMessage = 'test(state): implement widget integration tests for state';
-  final DateTime timestamp = DateTime.parse('2026-04-10T15:59:26+05:00');
+  final String commitRef = '299';
+  final String lastMessage = 'style(state): harmonize light and dark theme palette for state';
+  final DateTime timestamp = DateTime.parse('2026-04-10T20:14:48+05:00');
 }